- misc: another big change to the general API of libmedio. First,   removed some of the wrongfully placed __MEDIO_PRIVATE__ #ifdef   which caused third party programs to allocate not enough memory and   therefore run into certain memory issues. This however required to   implement another data hiding mechanism to the public classes of   libmedio such that no unnecessary methods and private methods are   exposed to the library user. There private classes were created   which embeed all private members of the main class and are publically   accessible by it. This also required to introduce default copy   constructor and assignment operator methods to all main libmedio   classes. Please note that also all copy constructors of header classes   were modified to NOT copy anything file related like the normally   corresponding MedIOData pointer. This in fact means that an "="   operation on two header object 'just' copies their content and   leave the data pointer intact.

+++ testcases/ecat/readwritematrix.cpp
@@ -293,7 +293,7 @@ int main(int argc, char* argv[])
 					break;
 				}
 
-				long j=0;
+				unsigned int j=0;
 				for(; j < MATRIX_SIZE; j++)
 				{
 					if(matrixData_frame2[j] != readBuf[j])
